Adjust single-click translation priority
