+++ docs/plan/y2026/m02-issue-twitter-show-more-bug/manifest__single-click-priority-over-modal_2026-03-14.md
@@ -1,0 +1,39 @@
+# Code Change Handoff Manifest
+
+## 1. Change Context
+
+- **Related Documents**:
+  - `src/1_content/README.md`
+- **Task Objectives**:
+  - This task attempted to adjust click-priority behavior when a new single-click word translation target appears underneath the visual zone of an existing translated word.
+  - The intended behavior was to allow the new word translation to take priority over opening the existing translation detail modal when the click can still resolve to a valid new word range.
+  - A follow-up adjustment attempted to reduce same-click race conditions between the content script's single-click translation path and the V2 translation hit-testing modal-open timer.
+
+## 2. File Change Audit
+
+| File Path | Change Type | Objective Description |
+| :--- | :--- | :--- |
+| `src/1_content/handlers/InputListener.ts` | Mod | Added synchronous same-event suppression of the V2 translation click handler when a higher-priority single-click word candidate is detected, and continued cancelling any pending modal-open timer before starting the new translation request. |
+| `src/1_content/handlers/utils/selectionValidator.ts` | Mod | Reworked single-click validation to resolve a candidate word range before rejecting the click as "inside active translation", intending to allow overlapped clicks to start a new translation when the detected word is not fully contained by an existing active translation. |
+| `src/1_content/handlers/utils/singleClickWordCandidate.ts` | Add | Added a shared helper for resolving a click-point word range and checking whether that range is fully contained by one existing translation, so validation and hit-testing can use the same priority rule. |
+| `src/1_content/ui/translationDisplayV2/hitTesting.ts` | Mod | Added logic intended to skip modal-open handling when a higher-priority single-click word candidate exists at the same click point, and added per-event / pending-timer suppression hooks to reduce double-trigger behavior. |
+
+## 3. AI Generation Disclaimer & Risk Warnings
+
+> **Important Note for Reviewer**:
+> The code in this submission was generated by an AI assistant based on documentation and local code inspection. **Do not assume the code logic is correct.** The implementation may contain hallucinations, flawed assumptions, incomplete understanding of browser event ordering, or non-ideal tradeoffs. Independent review is required.
+
+- [ ] **Logical Consistency**: Verify that a click on a new underlying word now consistently prefers starting that word's translation over opening the existing translation modal.
+- [ ] **Side Effects**: Verify that clicks on already translated text still open the intended modal when no higher-priority new word candidate is available.
+- [ ] **Edge Cases**: Verify behavior when the click lands on whitespace, punctuation, tooltip overlap zones, or partially occluded text where `tapWordDetector` may fail to resolve a stable word range.
+- [ ] **Propagation / Event-Order Semantics**: Verify the actual listener ordering between `InputListener.handleSingleClick` and `translationDisplayV2/hitTesting.handleClick`, and confirm the new same-event suppression does not silently break other click flows.
+- [ ] **Concurrency / Race Conditions**: Verify that the modal-open debounce timer cannot still fire after a new single-click translation has begun, especially under async validation or language-detection delays.
+- [ ] **State Synchronization**: Verify that the new helper's "fully contained by active translation" rule stays consistent with overlap cleanup and the current active range map when translations are added or removed quickly.
+- [ ] **Browser Compatibility**: Verify that repeated use of `caretRangeFromPoint` / click-point word resolution behaves consistently on Chrome pages with mixed inline elements and browser zoom.
+- [ ] **Detached DOM / SPA Lifecycle Behavior**: Verify that the new priority helper and same-event suppression do not create stale decisions when DOM layout shifts during SPA updates or immediately after line-height adjustments.
+
+## 4. Validation Notes
+
+- `npm run type-check` was run after the change.
+- `npm run build` was run after the change.
+- No automated tests were run for this task.
