+++ docs/prompt_files/review/code-review-expert/references/extension-architecture-checklist.md
